requirements.txt
python-dotenv
requests
nbtlib